Moved a create-exe test from the "run.rs" integration tests into "create_exe.rs"
